Loosened up restrictions some more
Option<T> is now a Semigroup and Monoid
if T is also a Semigroup and Monoid.

Option<T> only combines both T's if they are Some
otherwise it just returns the other Some value,
if both are None, it simply returns None.

This makes it so that an Option<T> where T is a Semigroup
is can now be seen as an optional value for example for configurations
and any none Option would be a requirement.