A simple formatting api for different styles.
When formatting a task we sometimes want to show id, have a colored output
or some other things depending on context.

Created a new trait to be able to call format that returns a String (type aliased).
This can be changed in the future.

Sometimes I want different looking output based on if I'm in interactive mode or CLI mode.