Add columns for formatting
Columns holds a prototype of a Cell, this way we can get the same formatting on each
column that correspond to the same column.

Cell can now be built with Default and change it's content with_content method.

SimpleFormatter now keeps the state of each column in memory.

The dispatcher in main had to change type so that we could capture some variables.